+++ src/translations/pl.ts
@@ -1,6 +1,8 @@
 /* Polish translations */
 
-export default {
+import { deepFreeze } from '../deep-freeze';
+
+export default deepFreeze({
   // Fallback Error Code
   unknown_error: 'Wystąpił nieoczekiwany błąd. Spróbuj ponownie później.',
   services: {
@@ -361,4 +363,4 @@ export default {
     // Functions-specific errors
     functions: {},
   },
-};
+});